refactor(mcp): extract completion/result_format/context leaf modules from server.rs (lab-kvji.24.1.1)
Move pure free fns + request-context inherent methods out of server.rs into
three new <500 LOC sibling modules. No behavior change — relocation + pub(crate)
visibility only (scope/admin gates widened to pub(crate) per Revision 2).

- completion.rs: action_schema, completion_info, complete_prompt_arg, service_name_completions
- result_format.rs: tool_error_envelope, hash_arguments, estimate_tokens*, format_dispatch_result, extract_error_info
- context.rs: request-context impl methods + auth/subject helpers + scope/admin gate fns
- logging.rs: repoint actor_key_from_extensions to mcp::context
- normalize_upstream_result intentionally NOT moved (consolidates into upstream.rs in .5)

server.rs 3492 -> 3147 LOC. Test name-set byte-identical to baseline (1423 pass, 24 skipped).
Bead tracking via commit msgs — remote dolt server down this session.
