+++ crates/lab/src/mcp/completion.rs
@@ -1,0 +1,64 @@
+//! Completion + per-service action-schema helpers for the MCP server.
+//!
+//! Pure free functions extracted from `server.rs` (bead `lab-kvji.24.1.1`).
+//! No behavior change — relocation + `pub(crate)` visibility only.
+
+use rmcp::model::CompletionInfo;
+use serde_json::Value;
+
+use crate::registry::ToolRegistry;
+
+/// JSON Schema for every service tool's input: `action` (required) + `params` (optional object).
+#[allow(clippy::expect_used)]
+pub(crate) fn action_schema() -> serde_json::Map<String, Value> {
+    serde_json::json!({
+        "type": "object",
+        "properties": {
+            "action": {
+                "type": "string",
+                "description": "Action to perform (e.g. \"movie.search\"). Use \"help\" to list all actions."
+            },
+            "params": {
+                "type": "object",
+                "description": "Action-specific parameters (varies per action)"
+            }
+        },
+        "required": ["action"]
+    })
+    .as_object()
+    .cloned()
+    .expect("schema literal is always an object")
+}
+
+pub(crate) fn completion_info(values: Vec<String>) -> CompletionInfo {
+    CompletionInfo {
+        total: Some(values.len() as u32),
+        has_more: Some(false),
+        values,
+    }
+}
+
+pub(crate) fn complete_prompt_arg(
+    registry: &ToolRegistry,
+    prompt_name: &str,
+    argument_name: &str,
+    prefix: &str,
+) -> CompletionInfo {
+    match (prompt_name, argument_name) {
+        ("run-action", "action") => completion_info(registry.action_name_completions(prefix)),
+        ("run-action" | "service-discover", "service") => {
+            completion_info(service_name_completions(registry, prefix))
+        }
+        _ => completion_info(Vec::new()),
+    }
+}
+
+pub(crate) fn service_name_completions(registry: &ToolRegistry, prefix: &str) -> Vec<String> {
+    registry
+        .services()
+        .iter()
+        .map(|service| service.name)
+        .filter(|name| name.starts_with(prefix))
+        .map(str::to_string)
+        .collect()
+}
